refactor(app): refactor AdditionalCustomLabwareSourceFolder
add AdditionalCustomLabwareSourceFolder and its test.

close RAUT-938 partially
