+++ app/src/organisms/AdvancedSettings/index.ts
@@ -1,3 +1,4 @@
+export * from './AdditionalCustomLabwareSourceFolder'
 export * from './ClearUnavailableRobots'
 export * from './EnableDevTools'
 export * from './OT2AdvancedSettings'
