+++ app/src/pages/AppSettings/AdvancedSettings.tsx
@@ -5,27 +5,17 @@ import { useSelector, useDispatch } from 'react-redux'
 import {
   ALIGN_CENTER,
   Box,
-  COLORS,
   DIRECTION_COLUMN,
   Flex,
-  Icon,
   JUSTIFY_SPACE_BETWEEN,
-  Link,
-  SPACING_AUTO,
   SPACING,
   TYPOGRAPHY,
 } from '@opentrons/components'
 
 import * as Config from '../../redux/config'
-import * as CustomLabware from '../../redux/custom-labware'
 import { SelectOption } from '../../atoms/SelectField/Select'
 import { SelectField } from '../../atoms/SelectField'
-import {
-  useTrackEvent,
-  ANALYTICS_CHANGE_CUSTOM_LABWARE_SOURCE_FOLDER,
-} from '../../redux/analytics'
 import { Divider } from '../../atoms/structure'
-import { TertiaryButton } from '../../atoms/buttons'
 import { StyledText } from '../../atoms/text'
 import {
   ClearUnavailableRobots,
@@ -36,18 +26,17 @@ import {
   ShowHeaterShakerAttachmentModal,
   U2EInformation,
   ShowLabwareOffsetSnippets,
+  AdditionalCustomLabwareSourceFolder,
 } from '../../organisms/AdvancedSettings'
 
 import type { Dispatch } from '../../redux/types'
 
 export function AdvancedSettings(): JSX.Element {
   const { t } = useTranslation(['app_settings', 'shared'])
-  const trackEvent = useTrackEvent()
   const channel = useSelector(Config.getUpdateChannel)
   const channelOptions: SelectOption[] = useSelector(
     Config.getUpdateChannelOptions
   )
-  const labwarePath = useSelector(CustomLabware.getCustomLabwareDirectory)
   const dispatch = useDispatch<Dispatch>()
   const handleChannel = (_: string, value: string): void => {
     dispatch(Config.updateConfigValue('update.channel', value))
@@ -105,69 +94,7 @@ export function AdvancedSettings(): JSX.Element {
           </Flex>
         </Flex>
         <Divider marginY={SPACING.spacing24} />
-        <Flex
-          alignItems={ALIGN_CENTER}
-          justifyContent={JUSTIFY_SPACE_BETWEEN}
-          gridGap={SPACING.spacing40}
-        >
-          <Flex flexDirection={DIRECTION_COLUMN}>
-            <StyledText
-              css={TYPOGRAPHY.h3SemiBold}
-              paddingBottom={SPACING.spacing8}
-              id="AdvancedSettings_customLabware"
-            >
-              {t('additional_labware_folder_title')}
-            </StyledText>
-            <StyledText as="p" paddingBottom={SPACING.spacing8}>
-              {t('additional_folder_description')}
-            </StyledText>
-            <StyledText
-              as="h6"
-              textTransform={TYPOGRAPHY.textTransformUppercase}
-              color={COLORS.grey50}
-              paddingBottom={SPACING.spacing4}
-            >
-              {t('additional_folder_location')}
-            </StyledText>
-            {labwarePath !== '' ? (
-              <Link
-                role="button"
-                css={TYPOGRAPHY.pRegular}
-                color={COLORS.black90}
-                onClick={() =>
-                  dispatch(CustomLabware.openCustomLabwareDirectory())
-                }
-                id="AdvancedSettings_sourceFolderLink"
-              >
-                {labwarePath}
-                <Icon
-                  height="0.75rem"
-                  marginLeft={SPACING.spacing8}
-                  name="open-in-new"
-                />
-              </Link>
-            ) : (
-              <StyledText as="p">{t('no_folder')}</StyledText>
-            )}
-          </Flex>
-          {
-            <TertiaryButton
-              marginLeft={SPACING_AUTO}
-              onClick={() => {
-                dispatch(CustomLabware.changeCustomLabwareDirectory())
-                trackEvent({
-                  name: ANALYTICS_CHANGE_CUSTOM_LABWARE_SOURCE_FOLDER,
-                  properties: {},
-                })
-              }}
-              id="AdvancedSettings_changeLabwareSource"
-            >
-              {labwarePath !== ''
-                ? t('change_folder_button')
-                : t('add_folder_button')}
-            </TertiaryButton>
-          }
-        </Flex>
+        <AdditionalCustomLabwareSourceFolder />
         <Divider marginY={SPACING.spacing24} />
         <PreventRobotCaching />
         <Divider marginY={SPACING.spacing24} />
