+++ app/src/pages/AppSettings/__test__/AdvancedSettings.test.tsx
@@ -1,26 +1,22 @@
 import * as React from 'react'
 import { MemoryRouter } from 'react-router-dom'
 import { resetAllWhenMocks } from 'jest-when'
-import { fireEvent, screen } from '@testing-library/react'
+import { screen } from '@testing-library/react'
 
 import { renderWithProviders } from '@opentrons/components'
 
 import { i18n } from '../../../i18n'
-import {
-  useTrackEvent,
-  ANALYTICS_CHANGE_CUSTOM_LABWARE_SOURCE_FOLDER,
-} from '../../../redux/analytics'
-import * as CustomLabware from '../../../redux/custom-labware'
 import * as Config from '../../../redux/config'
 import {
+  AdditionalCustomLabwareSourceFolder,
   ClearUnavailableRobots,
   EnableDevTools,
   OT2AdvancedSettings,
   OverridePathToPython,
   PreventRobotCaching,
+  ShowHeaterShakerAttachmentModal,
   ShowLabwareOffsetSnippets,
   U2EInformation,
-  ShowHeaterShakerAttachmentModal,
 } from '../../../organisms/AdvancedSettings'
 
 import { AdvancedSettings } from '../AdvancedSettings'
@@ -46,15 +42,8 @@ const render = (): ReturnType<typeof renderWithProviders> => {
   )
 }
 
-const getCustomLabwarePath = CustomLabware.getCustomLabwareDirectory as jest.MockedFunction<
-  typeof CustomLabware.getCustomLabwareDirectory
->
-const getChannelOptions = Config.getUpdateChannelOptions as jest.MockedFunction<
+const mockGetChannelOptions = Config.getUpdateChannelOptions as jest.MockedFunction<
   typeof Config.getUpdateChannelOptions
->
-
-const mockUseTrackEvent = useTrackEvent as jest.MockedFunction<
-  typeof useTrackEvent
 >
 const mockPreventRobotCaching = PreventRobotCaching as jest.MockedFunction<
   typeof PreventRobotCaching
@@ -81,15 +70,13 @@ const mockOverridePathToPython = OverridePathToPython as jest.MockedFunction<
 const mockShowHeaterShakerAttachmentModal = ShowHeaterShakerAttachmentModal as jest.MockedFunction<
   typeof ShowHeaterShakerAttachmentModal
 >
-
-let mockTrackEvent: jest.Mock
+const mockAdditionalCustomLabwareSourceFolder = AdditionalCustomLabwareSourceFolder as jest.MockedFunction<
+  typeof AdditionalCustomLabwareSourceFolder
+>
 
 describe('AdvancedSettings', () => {
   beforeEach(() => {
-    mockTrackEvent = jest.fn()
-    mockUseTrackEvent.mockReturnValue(mockTrackEvent)
-    getCustomLabwarePath.mockReturnValue('')
-    getChannelOptions.mockReturnValue([
+    mockGetChannelOptions.mockReturnValue([
       {
         label: 'Stable',
         value: 'latest',
@@ -113,6 +100,9 @@ describe('AdvancedSettings', () => {
     mockShowHeaterShakerAttachmentModal.mockReturnValue(
       <div>mock ShowHeaterShakerAttachmentModal</div>
     )
+    mockAdditionalCustomLabwareSourceFolder.mockReturnValue(
+      <div>mock AdditionalCustomLabwareSourceFolder</div>
+    )
   })
 
   afterEach(() => {
@@ -123,7 +113,6 @@ describe('AdvancedSettings', () => {
   it('renders correct titles', () => {
     const [{ getByText }] = render()
     getByText('Update Channel')
-    getByText('Additional Custom Labware Source Folder')
   })
 
   it('renders the update channel combobox and section', () => {
@@ -134,26 +123,11 @@ describe('AdvancedSettings', () => {
     getByRole('combobox', { name: '' })
   })
 
-  it('renders the custom labware section with source folder selected', () => {
-    getCustomLabwarePath.mockReturnValue('/mock/custom-labware-path')
-    const [{ getByText, getByRole }] = render()
-    getByText(
-      'If you want to specify a folder to manually manage Custom Labware files, you can add the directory here.'
-    )
-    getByText('Additional Source Folder')
-    getByRole('button', { name: 'Change labware source folder' })
+  it('should render mock OT-2 Advanced Settings Tip Length Calibration Method section', () => {
+    render()
+    screen.getByText('mock AdditionalCustomLabwareSourceFolder')
   })
 
-  it('renders the custom labware section with no source folder selected', () => {
-    const [{ getByText, getByRole }] = render()
-    getByText('No additional source folder specified')
-    const btn = getByRole('button', { name: 'Add labware source folder' })
-    fireEvent.click(btn)
-    expect(mockTrackEvent).toHaveBeenCalledWith({
-      name: ANALYTICS_CHANGE_CUSTOM_LABWARE_SOURCE_FOLDER,
-      properties: {},
-    })
-  })
   it('should render mock OT-2 Advanced Settings Tip Length Calibration Method section', () => {
     render()
     screen.getByText('mock OT2AdvancedSettings')
